ADD: datasources exposed to retrieve params
To run a pipeline or a release, the param id's are required.
It becomes a difficult job to trigger a pipeline or release by maintaining and tracking resource ids.

This feature set brings up a set of data sources which can be used to trigger releases and pipelin runs.

A new addon is also injected to make releases simpler.